move search from es to elastic package
