--- es/faces_index.go
@@ -1,4 +1,0 @@
-package es
-
-// FaceIndexName is the name of the index of the faces in elasticsearch
-var FaceIndexName string = "faces"
